Add some Master utils, TestPassed

+++ src/Git/git_pull.jl
@@ -4,18 +4,22 @@
     before returning.
     Return true if no error was raised
 """
-function git_force_pull(tries = 5, maxwt = 5)
+function git_pull(tries = 5, maxwt = 5; force = false)
     try
         # TODO: Use LibGit2
-        run(`git fetch`)
-        run(`git reset --hard FETCH_HEAD`)
+        if force
+            run(`git fetch`)
+            run(`git reset --hard FETCH_HEAD`)
+        else
+            run(`git pull`)
+        end
         log("Local repo updated") # TODO: regulate frec
     catch err
         log(err)
         err isa InterruptException && rethrow(err)
         sleep(maxwt * rand())
         tries <= 0 && return false
-        git_force_pull(tries - 1, maxwt)
+        git_pull(tries - 1, maxwt; force = force)
     end
     return true
 end
